requirements.txt
aiomysql
fastapi
uvicorn
tortoise-orm
pydantic
